+++ bin/dialogs/preferencespanels/tagspreviewpanel.py
@@ -90,9 +90,12 @@ class TagsPreviewPanel(wx.Panel):
         self.Layout()
         self.Update()
 
-    def DoRefresh(self):
+    def DoRefresh(self, event=None):
         TagsSelected = self.TagDropdown.GetValue()
         self.BuildTagsList(TagsSelected)
+
+    def reloadFromSettings(self):
+        self.DoRefresh()
 
     def BuildTagsList(self, TagsSelected):
         self.TagsList.DeleteAllItems()